Move Block Hooks logic to a new trait
